+++ flake.nix
@@ -27,6 +27,8 @@
                 cargo
                 clippy
                 glibcLocales
+                go
+                jq
                 kind
                 kubernetes-helm
                 patchutils