+++ src/features/admission_plugins.rs
@@ -20,6 +20,7 @@ use kube::CustomResourceExt;
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, ClusterFeatureValues)]
+#[allow(dead_code)]
 pub struct FeatureValues {
     #[serde(rename = "admissionControlList")]
     pub admission_control_list: String,